add Readme file

+++ Exercice1/src/main/java/com/javaIntellij/TropViteException.java
@@ -1,7 +1,7 @@
 package com.javaIntellij;
 
 public class TropViteException extends Exception{
-    public TropViteException(int vitesse) {
+    public TropViteException(double vitesse) {
         super("C'est une exception de type TropViteException. Vitesse en cause : " + vitesse);
     }
 }
