+++ Exercice3/src/main/java/com/javaIntellij/NoteInvalideException.java
@@ -1,7 +1,7 @@
 package com.javaIntellij;
 
 public class NoteInvalideException extends Exception{
-    public NoteInvalideException(int valeur) {
+    public NoteInvalideException(double valeur) {
         super("Exception de type NoteInvalideException. Note invalide : "+ valeur);
     }
 }
